Make configuration more dynamic
Treat the config subdir as the desired state of the user homedir.
This means that the file structure of config will be replicated in
the user home - the playbook will ensure the whole directory tree is
replicated and every regular file is symlinked from user home to its
location in the dotfile repo.

I believe this approach has some advantages:
- It is easier to add new configuration. Just put it under config/ as if
  it was your home dir (for example put you vim conf under
  config/.config/nvim/init.vim)
- At some point I may replace the config subdir with a separae git repo.
  This way another person could just clone their dotfiles there and it
  will be linked assuming that the proper structure is maintained